% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudhsmv2_operations.R
\name{delete_backup}
\alias{delete_backup}
\title{Deletes a specified AWS CloudHSM backup}
\usage{
delete_backup(
  BackupId = "string"
)
}
\arguments{
\item{BackupId}{[required] The ID of the backup to be deleted. To find the ID of a backup, use the DescribeBackups operation.}
}
\description{
Deletes a specified AWS CloudHSM backup. A backup can be restored up to 7 days after the DeleteBackup request. For more information on restoring a backup, see RestoreBackup
}
\examples{

}
