% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudhsmv2_operations.R
\name{delete_cluster}
\alias{delete_cluster}
\title{Deletes the specified AWS CloudHSM cluster}
\usage{
delete_cluster(ClusterId)
}
\arguments{
\item{ClusterId}{The identifier (ID) of the cluster that you are deleting. To find the cluster ID, use DescribeClusters.}
}
\description{
Deletes the specified AWS CloudHSM cluster. Before you can delete a cluster, you must delete all HSMs in the cluster. To see if the cluster contains any HSMs, use DescribeClusters. To delete an HSM, use DeleteHsm.
}
\examples{

}
