% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudhsmv2_operations.R
\name{describe_clusters}
\alias{describe_clusters}
\title{Gets information about AWS CloudHSM clusters}
\usage{
describe_clusters(Filters = NULL, NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{Filters}{One or more filters to limit the items returned in the response.

Use the \code{clusterIds} filter to return only the specified clusters. Specify clusters by their cluster identifier (ID).

Use the \code{vpcIds} filter to return only the clusters in the specified virtual private clouds (VPCs). Specify VPCs by their VPC identifier (ID).

Use the \code{states} filter to return only clusters that match the specified state.}

\item{NextToken}{The \code{NextToken} value that you received in the previous response. Use this value to get more clusters.}

\item{MaxResults}{The maximum number of clusters to return in the response. When there are more clusters than the number you specify, the response contains a \code{NextToken} value.}
}
\description{
Gets information about AWS CloudHSM clusters.
}
\details{
This is a paginated operation, which means that each response might contain only a subset of all the clusters. When the response contains only a subset of clusters, it includes a \code{NextToken} value. Use this value in a subsequent \code{DescribeClusters} request to get more clusters. When you receive a response with no \code{NextToken} (or an empty or null value), that means there are no more clusters to get.
}
\section{Accepted Parameters}{
\preformatted{describe_clusters(
  Filters = list(
    list(
      "string"
    )
  ),
  NextToken = "string",
  MaxResults = 123
)
}
}

