% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudhsmv2_package.R
\docType{package}
\name{paws.cloudhsmv2-package}
\alias{paws.cloudhsmv2}
\alias{paws.cloudhsmv2-package}
\title{paws.cloudhsmv2: AWS CloudHSM V2}
\description{
For more information about AWS CloudHSM, see [AWS
CloudHSM](http://aws.amazon.com/cloudhsm/) and the [AWS CloudHSM User
Guide](http://docs.aws.amazon.com/cloudhsm/latest/userguide/).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
