% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudhsmv2_operations.R
\name{tag_resource}
\alias{tag_resource}
\title{Adds or overwrites one or more tags for the specified AWS CloudHSM cluster}
\usage{
tag_resource(ResourceId, TagList)
}
\arguments{
\item{ResourceId}{[required] The cluster identifier (ID) for the cluster that you are tagging. To find the cluster ID, use DescribeClusters.}

\item{TagList}{[required] A list of one or more tags.}
}
\description{
Adds or overwrites one or more tags for the specified AWS CloudHSM cluster.
}
\section{Accepted Parameters}{
\preformatted{tag_resource(
  ResourceId = "string",
  TagList = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

